1) RawMaterial new field deleted to remove unused elements. 2) User_Controller added hidden filter on deleted column for all users.

+++ build/templates/controllers/Production_Controller_php.xsl
@@ -1041,7 +1041,9 @@ UPDATE public.production_sites
 				//throw $e;
 				
 			}finally{
-				unlink($SYGNAL_FILE);
+				if(file_exists($SYGNAL_FILE)){
+					unlink($SYGNAL_FILE);
+				}
 			}			
 		}
 	}
